:package: feat: Arrange code in MVC
