+++ static/upload.js
@@ -24,7 +24,7 @@ uploadVideoButton.addEventListener("click", () => {
   const theFile = video.files[0];
 
   fileReader.onload = async (ev) => {
-    const CHUNK_SIZE = 50000;
+    const CHUNK_SIZE = 5000000;
     const chunkCount = parseInt(ev.target.result.byteLength / CHUNK_SIZE);
     console.log(chunkCount);
     console.log("Read successfully");
