+++ structs/video.go
@@ -1,0 +1,7 @@
+package structs
+
+type Video struct {
+	ID          string `json:"id"`
+	Title       string `json:"title"`
+	Description string `json:"description"`
+}
